This change removes CG from O3D and uses GLSL instead of HLSL. The current "hack" for shaders is that the shader string you supply to Effect::LoadFromFXString requires the lines
// #o3d SplitMarker                                                \n"

Above which is the vertex shader and below which
is the fragment shader.

Since GLSL has no semantics the names are used
for semantics so to get an automatically supplied
worldViewProjection you must name the uniform
worldViewProjection as in

uniform mat4 worldViewProjection

Similarly for attributes (vertex streams) you must
name them to match O3D semantics as in

attribute vec4 texcoord1
attribute vec3 normal

etc.

Included in this is one working sample in
o3d/samples/hellocube-glsl.html

Other things: Added a glsl field to clientInfo
so that a program can check if it needs to supply
GLSL or HLSL. eg client.clientInfo.glsl == true

Updated the O3D libraries to store that info
in o3djs.base.glsl so that the libaries can be
modified to supply GLSL if we want.

Note: To run this on windows you must copy glew32.dll
and glut32.dll to <AppData>...\Mozilla\plugins

To get this to actually work in GLES2 will
require 

(*) renaming a few functions
(*) telling O3D there is no NPOT support or if possible
    no NPOT support unless the texture has only 1 level.
(*) removing LargeGeometry and FloatingPointTexture support.
(*) Making it link with the GLES2 code


Review URL: http://codereview.chromium.org/527028

git-svn-id: http://src.chromium.org/svn/trunk/src/o3d/samples/o3djs@35845 4ff67af0-8c30-449e-8e8b-ad334ec8d88c